converter-service/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
cloudinary==1.40.0
dnspython==2.6.1
ffmpeg==1.4
idna==3.7
pika==1.3.2
pydub==0.25.1
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.1
